Haptics challenge done
Haptics are hapticing (mostly.) Issues: when I pick up an object that should not do the long pulse, it does the long pulse. Also if i pick it up with the left hand the long pulse is always on the right side. The cylinder you cannot pick up at all right now either?